Initial commit of actual code
The code already parses an Etim file to get the translations for the
codes and other stuff.
It also parses a Bmecat file to Products.
Also it can already build a csv file from the parsed products.
The csv file is not complete yet, though.
The properties (features) of the products are not yet used.

To put the properties into the csv file they have to exist in gambios
database.
So there is also a DbAccessor which is supposed to create all the
necessary properties in gambios database by looking at all of the
products features.
Only a part of the DbAccessor is done yet.